First commit with modified Villager behavior
This is just a test to get Villagers to toggle follow/stay on-right-click

PathfinderGoalFollowPlayer.java is a modified version of
PathfinderGoalFollowOwner

Hostage.java contains the modified Villager behavior, which clears all the
default movements and adds FollowOwner movement.

No commands.

config.yml is not implemented yet.

ExtractionTimer.java is not implemented yet.

HostageArena.java is the Listener class.