Added VictoryCondition onHostageDeath()
onHostageDeath(EntityDeathEvent e) :

if one of the teams kills 2 hostages, then the match will end.

To-Do:

Add a victory condition that checks if a team has brought 1 or more
hostages to the extraction zone.

Current design: onStart() creates a HostageTracker() that checks on
hostages every second. Need to add a distance check, and finalize the
parameters for the constructor.

A successful distance check will activate the ExtractionTimer().