Deleted HostageListener.
The purpose of this class was to add behavior to hostages spawned outside
of arenas with the '/vips spawn' command.

The original idea was to have a Listener for each hostage that would be
registered on '/vips spawn' and unregistered when the hostage died.

Was supposed to be used for testing purposes.

Alternatively, we could have had ONE Listener that detected the entity
type, and conditional reacted to Hostages only. However, I didn't want to
run into the problem of having this listener cause problems in the Arena:
i.e. Cause double dmg to hostages (because the arena listener would be one
source of dmg, and the HostageListener would be the second source of dmg).