Added ExtractionZone for future use.
The ExtractionZone class will be a different implementation for detecting
if hostages are at the rescue point.

The first implementation is a distance calculation from the Hostage to the
ExtractionPoint along with a comparison of that distance to the radius of
the extraction zone. However, the distance calculation requires an
expensive square root calculation.

That's why I want to prepare for other, more efficient implementations.

The ExtractionZone class will merely be a bounding box implementation. So
that the conditional check for (is a Hostage in the zone ?) will merely be
a comparison of the Hostage's X, Y, Z with that of the bounding box.

Similar to how Alkarin does it for ArenaParkour.