Refactored HostageTracker
HostageTracker now asks for most of what it needs in its constructor:
What match ?
What Set of Hostages ?
What are the Extraction Points ?

However, it supplies the plugin field statically from thin air:
Bukkit.getPluginManager().getPlugin("HostageArena");

It also instantiates its own ExtractionTimer() object rather than asking
for one. (In the future, i might make a different ExtractionTimer for each
ExtractionPoint... this is probably how it should be).

The job of HostageTracker:

HostageTracker is run once per second.

It's job is to ask:
'Are there any hostages in the extraction zone ?'
- if the timer hasn't started yet, then start it.
- if the timer has started, then tell the timer which hostages are in the
  extraction zone.