Adds TODO, Fixes 1.8 NPC event bug, Fixes/Adds etc
TODO:
test addArenaListener(): I suspect this is functioning properly and that
any bugs in the past where most likely due to an incorrect annotation, or
due to being passed a Listener instead of an ArenaListener.

TODO:
Finish the game rules inside MatchResultEvent, and handle ties. As it
stands right now, either team can rescue the hostages. Which means that
the VIPs should spawn in the middle of the arena, with an extraction zone
for each team, and the teams should fight over the hostages. This could
become a separate game rule if admins desire this setup. But conventional
rules have one team as the guards, and the other team is the rescuers. If
the rescuers do not kill the guards (terrorists), or accomplish their
objective (rescue the hostages), then the guards should win when time
expires. This is un-implemented.

Fixes 1.8 NPC bug that prevented Hostages from following:
Minecraft 1.8 added a new event: PlayerInteractAtEntityEvent. The problem
is for plugins that listen to PlayerInteractEntityEvent because they're
receiving that event... plus the new one, PlayerInteractAtEntityEvent.
Which means that it's being called twice: The first call would assign an
owner and the hostage would follow the owner for nanoseconds. The second
call would remove the owner. This created NPCs that weren't following bug.

The solution is to check event.getEventName() and if it doesn't equal
"PlayerInteractEntityEvent" then just exit out of the method.

Fixes player death bug:
The bug occurs when a player dies while controlling a hostage... The owner
of the hostage didn't get cleared, so the Hostage would run insanely fast
to the owner when he spawned (at light-speed in the blink of an eye).
But it's an easy fix to listen for the PlayerDeathEvent and check all the
VIPs to see if their owner died.

FIRES NEW EVENT:
Now fires the new HostageDeathEvent !